+++ app/src/main/res/layout/fragment_near.xml
@@ -9,9 +9,4 @@
         android:layout_height="50dp">
 
     </android.support.v7.widget.Toolbar>
-    <com.baidu.mapapi.map.MapView
-        android:id="@+id/bmapView"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:clickable="true" />
 </LinearLayout>
